+++ lib/redmine_messenger/hooks.rb
@@ -18,8 +18,10 @@ module RedmineMessenger
 
       repository = changeset.repository
 
-      if Setting.host_name.to_s =~ /\A(https?\:\/\/)?(.+?)(\:(\d+))?(\/.+)?\z/i
-        host, port, prefix = $2, $4, $5
+      if Setting.host_name.to_s =~ %r{/\A(https?\:\/\/)?(.+?)(\:(\d+))?(\/.+)?\z/i}
+        host = Regexp.last_match(2)
+        port = Regexp.last_match(4)
+        prefix = Regexp.last_match(5)
         revision_url = Rails.application.routes.url_for(
           controller: 'repositories',
           action: 'revision',